✨ Feat:	재사용을 위한 followButton 분리
